pfcp_test: add missing 'pdi_present = true'
test_enc_dec[8] includes a PDI IE in Update PDR. The PDI is optional
here, so it requires setting a presence flag to true. This was missing,
hence the PDI part was not being tested at all. Add this presence flag.

This was probably a copy-paste error from testing the Create PDR IE,
where the PDI is mandatory and has no presence flag.

Change-Id: Idc70ae23f15d3af65f98557cc51a111f2e6d75c9